ll_merge
