+++ data_structures/linked_list/README.md
@@ -1,0 +1,21 @@
+# Linked List
+
+
+# Features
+- [X]Create a Class for a Node which is aware of the value as val and next as _next
+- [X]Ensure that you have a __str__ method defined to return the value of the node when printed
+- [ ]Create a Class for a LinkedList which creates an empty Linked List when instantiated.
+- [ ]This class should be aware of a default None value assigned to head when the list is created.
+- [ ]This class should be aware of the len of the list, which represents the count of Nodes in the list at any time
+- [ ]This class should have the ability to accept an iterable as an argument when instantiated, such as [1, 2, 3, 4], and creates a new - [ ]Node in the list for each value in the iterable.
+- [ ]Define any further magic methods such as len and str to support user functionality and informative responses.
+- [ ]Define a method called insert which takes any value as an argument and adds that value to the head of the list with an O(1) Time performance.
+- [ ]Define a method called find which takes any value as an argument and returns True or False depending on whether that value exists as a Node value within the list.
+- [ ]At no time should an exception or stack trace be show to the end user. Catch and handle any such exceptions and return a printed value or operation which cleanly represents the state and either Stops execution cleanly or provides the user with clear direction and output.
+
+# Testing
+- [ ]Every bit of functionality that you add should be tested.
+- [ ]As a general rule at this time, you should have a test for valid, invalid, and edge case variants for every function that you define. There are exceptions. The exceptions are not the rule.
+
+# Change Log
+-
